requirements.txt
streamlit
ffmpeg-python
pydub
SpeechRecognition
deep-translator
gtts
numpy
requests
